Fixed bugs

+++ wp-content/themes/player-key/inc/api/PKI_REST_Entities_Controller.php
@@ -2,11 +2,15 @@
 
 namespace api;
 
+use Cassandra\Numeric;
 use Couchbase\User;
 use WP_REST_Controller;
 use WP_REST_Request;
 
 class PKI_REST_Entities_Controller extends WP_REST_Controller {
+	const WEEK_IN_SECONDS = 604800;
+	const DAY_IN_SECONDS = 86400;
+
 	function __construct() {
 		$this->namespace = 'pki/v1';
 		$this->rest_base = 'entities';
@@ -112,13 +116,13 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 	}
 
 	function edit_parent( WP_REST_Request $request ) {
-		$data    = json_decode( $request->get_body(), true );
-		$user_id = get_option( $data['token'] );
+		$data     = json_decode( $request->get_body(), true );
+		$coach_id = get_option( $data['token'] );
 
-		if ( ! empty( $user_id ) ) {
-			$user = get_user_by( 'ID', $user_id );
+		if ( ! empty( $coach_id ) ) {
+			$coach = get_user_by( 'ID', $coach_id );
 
-			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_team' ) ) {
+			if ( ! is_wp_error( $coach ) && $coach !== false && user_can( $coach_id, 'create_team' ) ) {
 
 				$parent_id = wp_update_user( [
 					'ID'         => $data['parentId'],
@@ -129,29 +133,10 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 				] );
 
 				if ( ! is_wp_error( $parent_id ) ) {
-					$parents = get_users( [
-						'role'       => 'parent',
-						'meta_query' => [
-							[
-								'key'   => 'coach',
-								'value' => $user_id,
-							]
-						],
-					] );
+					$parents = $this->get_parents_list( $coach_id );
 
 					if ( ! empty( $parents ) ) {
-						$parents_data = [];
-						foreach ( $parents as $parent ) {
-							$parents_data[] = [
-								'ID'         => $parent->ID,
-								'first_name' => $parent->first_name,
-								'last_name'  => $parent->last_name,
-								'email'      => $parent->user_email,
-								'login'      => $parent->user_login,
-							];
-						}
-
-						wp_send_json_success( $parents_data );
+						wp_send_json_success( $parents );
 					}
 				}
 
@@ -204,7 +189,7 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 			$user = get_user_by( 'ID', $user_id );
 
 			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_athlete' ) ) {
-				$coach = get_field( 'coach', $data['athleteId'] );
+				$coach  = get_field( 'coach', $data['athleteId'] );
 				$parent = get_field( 'parent', $data['athleteId'] );
 				if ( $coach->ID == $user_id || $parent->ID == $user_id ) {
 
@@ -220,7 +205,7 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 						update_field( 'parent', $data['form']['parent']['code'], $athlete_id );
 						update_field( 'team', $data['form']['team']['code'], $athlete_id );
 
-						if ($data['currentRole'] === 'parent') {
+						if ( $data['currentRole'] === 'parent' ) {
 							$team_id = $data['form']['team']['code'];
 							$coach   = get_field( 'coach', $team_id );
 							update_field( 'coach', $coach, $athlete_id );
@@ -274,8 +259,14 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_team' ) ) {
 
 				if ( get_field( 'coach', $data['teamId'] )->ID == $user_id ) {
-					wp_delete_post( $data['teamId'] );
-					wp_send_json_success();
+
+					$athletes = $this->get_athletes_list( 'coach', $user_id );
+					if ( count( $athletes ) === 0 ) {
+						wp_delete_post( $data['teamId'] );
+						wp_send_json_success();
+					} else {
+						wp_send_json_error( 'You cannot delete a team while it has athletes.' );
+					}
 				}
 			}
 		}
@@ -283,17 +274,57 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 	}
 
 	function remove_parent( WP_REST_Request $request ) {
-		$data    = json_decode( $request->get_body(), true );
-		$user_id = get_option( $data['token'] );
+		$data      = json_decode( $request->get_body(), true );
+		$coach_id  = get_option( $data['token'] );
+		$parent_id = $data['parentId'];
 
-		if ( ! empty( $user_id ) ) {
-			$user = get_user_by( 'ID', $user_id );
+		if ( ! empty( $coach_id ) ) {
+			$coach = get_user_by( 'ID', $coach_id );
 
-			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_team' ) ) {
+			if ( ! is_wp_error( $coach ) && $coach !== false && user_can( $coach_id, 'create_team' ) ) {
 
-				if ( get_field( 'coach', 'user_' . $data['parentId'] ) == $user_id ) {
-					require_once( ABSPATH . 'wp-admin/includes/user.php' );
-					wp_delete_user( $data['parentId'] );
+				$parent_coaches = get_field( 'coaches', 'user_' . $parent_id );
+				$parent_coaches = array_map( function ( $coach ) {
+					return $coach->ID;
+				}, $parent_coaches );
+
+				if ( in_array( $coach_id, $parent_coaches ) ) {
+
+					$athletes = get_posts( [
+						'numberposts' => - 1,
+						'post_type'   => 'athlete',
+						'meta_query'  => [
+							'relation' => 'AND',
+							[
+								'key'   => 'coach',
+								'value' => $coach_id,
+							],
+							[
+								'key'   => 'parent',
+								'value' => $parent_id,
+							]
+						]
+					] );
+
+					if ( ! empty( $athletes ) ) {
+						foreach ( $athletes as $athlete ) {
+							wp_delete_post( $athlete->ID );
+						}
+					}
+
+					$key = array_search( $coach_id, $parent_coaches );
+					unset( $parent_coaches[ $key ] );
+					update_field( 'coaches', $parent_coaches, 'user_' . $parent_id );
+
+//					if ( count( $parent_coaches ) === 1 && get_user_meta( $parent_id, 'initial_coach', true ) === $coach_id ) {
+//						require_once( ABSPATH . 'wp-admin/includes/user.php' );
+//						wp_delete_user( $data['parentId'] );
+//					} else {
+//						if ( ( $key = array_search( $coach_id, $parent_coaches ) ) !== false ) {
+//							unset( $parent_coaches[ $key ] );
+//							update_field('coaches', $parent_coaches, 'user_' . $parent_id);
+//						}
+//					}
 					wp_send_json_success();
 				}
 			}
@@ -309,7 +340,7 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 			$user = get_user_by( 'ID', $user_id );
 			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_athlete' ) ) {
 
-				$coach = get_field( 'coach', $data['athleteId'] );
+				$coach  = get_field( 'coach', $data['athleteId'] );
 				$parent = get_field( 'parent', $data['athleteId'] );
 				if ( $coach->ID == $user_id || $parent->ID == $user_id ) {
 					wp_delete_post( $data['athleteId'] );
@@ -321,13 +352,13 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 	}
 
 	function create_parent( WP_REST_Request $request ) {
-		$data    = json_decode( $request->get_body(), true );
-		$user_id = get_option( $data['token'] );
+		$data     = json_decode( $request->get_body(), true );
+		$coach_id = get_option( $data['token'] );
 
-		if ( ! empty( $user_id ) ) {
-			$user = get_user_by( 'ID', $user_id );
+		if ( ! empty( $coach_id ) ) {
+			$coach = get_user_by( 'ID', $coach_id );
 
-			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_team' ) ) {
+			if ( ! is_wp_error( $coach ) && $coach !== false && user_can( $coach_id, 'create_team' ) ) {
 
 				$parent_id = wp_insert_user( [
 					'user_login' => $data['form']['login'],
@@ -339,46 +370,29 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 				] );
 
 				if ( ! is_wp_error( $parent_id ) ) {
-					update_field( 'coach', $user_id, 'user_' . $parent_id );
+					add_user_meta( $parent_id, 'initial_coach', $coach_id );
+					update_field( 'coaches', [ $coach_id ], 'user_' . $parent_id );
 					update_field( 'is_activated', 'no', 'user_' . $parent_id );
 
-					$token = wp_hash( $parent_id . $data['form']['firstName'] . $data['form']['lastName'] . $data['form']['email'] );
-					add_user_meta( $parent_id, 'activation_token', $token, true );
-
-					$url     = get_site_url() . '/activation/parent/' . '?id=' . $parent_id . '&token=' . $token;
-					$message = file_get_contents( TEMPLATE_DIR . '/inc/templates/emails/parent-activation-email.php' );
-					$message = str_replace( array( '{{url}}', '{{coach}}' ), array(
+					$activation_token = wp_hash( $parent_id . $data['form']['firstName'] . $data['form']['lastName'] . $data['form']['email'] );
+					$url              = get_site_url() . '/activation/parent/' . '?id=' . $parent_id . '&token=' . $activation_token;
+					$message          = file_get_contents( TEMPLATE_DIR . '/inc/templates/emails/parent-activation-email.php' );
+					$message          = str_replace( array( '{{url}}', '{{coach}}' ), array(
 						$url,
-						$user->first_name . ' ' . $user->last_name
+						$coach->first_name . ' ' . $coach->last_name
 					), $message );
 
-					$is_mail_sent = wp_mail( $data['form']['email'], 'Activate your account on PlayerKey ID', $message, [
+					wp_mail( $data['form']['email'], 'Activate your account on PlayerKey ID', $message, [
 						'content-type: text/html',
 					] );
 
-					$parents = get_users( [
-						'role'       => 'parent',
-						'meta_query' => [
-							[
-								'key'   => 'coach',
-								'value' => $user_id,
-							]
-						],
-					] );
+					add_user_meta( $parent_id, 'activation_token', $activation_token, true );
+					add_option( $activation_token, time() );
+
+					$parents = $this->get_parents_list( $coach_id );
 
 					if ( ! empty( $parents ) ) {
-						$parents_data = [];
-						foreach ( $parents as $parent ) {
-							$parents_data[] = [
-								'ID'         => $parent->ID,
-								'first_name' => $parent->first_name,
-								'last_name'  => $parent->last_name,
-								'email'      => $parent->user_email,
-								'login'      => $parent->user_login,
-							];
-						}
-
-						wp_send_json_success( $parents_data );
+						wp_send_json_success( $parents );
 					}
 				}
 
@@ -427,13 +441,13 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 					} else if ( $data['currentRole'] === 'parent' ) {
 						$team_id = $data['form']['team']['code'];
 						// Team coach
-						$coach   = get_field( 'coach', $team_id );
+						$coach = get_field( 'coach', $team_id );
 
 						//Parent coaches (can be multiple)
-						$parent_coaches = get_field( 'coach', 'user_' . $user_id );
+						$parent_coaches   = get_field( 'coaches', 'user_' . $user->ID, true );
 						$parent_coaches[] = $coach;
 
-						update_field( 'coach', $parent_coaches, 'user_' . $user_id );
+						update_field( 'coaches', $parent_coaches, 'user_' . $user->ID );
 						update_field( 'coach', $coach, $athlete_id );
 
 						$args['meta_query'] = [
@@ -508,45 +522,87 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 	}
 
 	function get_parents( WP_REST_Request $request ) {
-		$data    = json_decode( $request->get_body(), true );
-		$user_id = get_option( $data['token'] );
+		$data     = json_decode( $request->get_body(), true );
+		$coach_id = get_option( $data['token'] );
 
-		if ( ! empty( $user_id ) ) {
-			$user = get_user_by( 'ID', $user_id );
+		if ( ! empty( $coach_id ) ) {
+			$coach = get_user_by( 'ID', $coach_id );
 
-			if ( ! is_wp_error( $user ) && $user !== false && user_can( $user_id, 'create_team' ) ) {
-				$parents = get_users( [
-					'role'       => 'parent',
-					'meta_query' => [
-						[
-							'key'   => 'coach',
-							'value' => $user_id,
-							'compare'   => 'LIKE'
-						]
-					],
-				] );
+			if ( ! is_wp_error( $coach ) && $coach !== false && user_can( $coach_id, 'create_team' ) ) {
+
+				$parents = $this->get_parents_list( $coach_id );
 
 				if ( ! empty( $parents ) ) {
-					$parents_data = [];
-					foreach ( $parents as $parent ) {
-						$parents_data[] = [
-							'ID'           => $parent->ID,
-							'first_name'   => $parent->first_name,
-							'last_name'    => $parent->last_name,
-							'email'        => $parent->user_email,
-							'login'        => $parent->user_login,
-							'is_activated' => get_field( 'is_activated', 'user_' . $parent->ID ),
-						];
-					}
-
-					wp_send_json_success( $parents_data );
+					wp_send_json_success( $parents );
 				}
 
-				wp_send_json_error( 'Teams not found' );
+				wp_send_json_error( 'Parents not found' );
 			}
 
 			wp_send_json_error( 'There are not enough permissions to perform this action' );
 		}
+	}
+
+	/**
+	 * @param int $coach_id
+	 *
+	 * @return array
+	 */
+	private function get_parents_list( int $coach_id ): array {
+		$parents_data = [];
+		$parents      = get_users( [
+			'role'       => 'parent',
+			'orderby'    => 'display_name',
+			'meta_query' => [
+				[
+					'key'     => 'coaches',
+					'value'   => $coach_id,
+					'compare' => 'LIKE'
+				]
+			],
+		] );
+
+		if ( ! empty( $parents ) ) {
+			foreach ( $parents as $parent ) {
+				$is_activated = get_field( 'is_activated', 'user_' . $parent->ID );
+				if ( $is_activated === 'no' ) {
+					$activation_token = get_user_meta( $parent->ID, 'activation_token', true );
+
+					if ( ! empty( $activation_token ) && time() - (int) get_option( $activation_token, true ) > self::DAY_IN_SECONDS ) {
+						$is_activated = 'expired';
+						update_field( 'is_activated', $is_activated, 'user_' . $parent->ID );
+					}
+				}
+
+				$athletes = get_posts( [
+					'numberposts' => - 1,
+					'post_type'   => 'athlete',
+					'meta_query'  => [
+						'relation' => 'AND',
+						[
+							'key'   => 'coach',
+							'value' => $coach_id,
+						],
+						[
+							'key'   => 'parent',
+							'value' => $parent->ID,
+						]
+					]
+				] );
+
+				$parents_data[] = [
+					'ID'           => $parent->ID,
+					'first_name'   => $parent->first_name,
+					'last_name'    => $parent->last_name,
+					'email'        => $parent->user_email,
+					'login'        => $parent->user_login,
+					'is_activated' => $is_activated,
+					'athletes'     => count( $athletes ),
+				];
+			}
+		}
+
+		return $parents_data;
 	}
 
 	function get_teams( WP_REST_Request $request ) {
@@ -556,7 +612,7 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 		if ( ! empty( $user_id ) ) {
 			$user = get_user_by( 'ID', $user_id );
 
-			if ( ! is_wp_error( $user ) && $user !== false && ( user_can( $user_id, 'create_team' ) || user_can( $user_id, 'create_athlete' ) ) ) {
+			if ( $user !== false && ! is_wp_error( $user ) && ( user_can( $user_id, 'create_team' ) || user_can( $user_id, 'create_athlete' ) ) ) {
 				$args = [
 					'numberposts' => - 1,
 					'post_type'   => 'team',
@@ -573,7 +629,16 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 
 				$teams = get_posts( $args );
 				if ( ! empty( $teams ) ) {
-					wp_send_json_success( $teams );
+					$teams_data = [];
+					foreach ( $teams as $team ) {
+						$athletes     = $this->get_athletes_list( 'coach', $user_id );
+						$teams_data[] = [
+							'ID'         => $team->ID,
+							'post_title' => $team->post_title,
+							'athletes'   => count( $athletes ),
+						];
+					}
+					wp_send_json_success( $teams_data );
 				}
 
 				wp_send_json_error( 'Teams not found' );
@@ -581,6 +646,46 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 
 			wp_send_json_error( 'There are not enough permissions to perform this action' );
 		}
+	}
+
+	/**
+	 * @param string $role
+	 * @param int $id
+	 *
+	 * @return array
+	 */
+	private function get_athletes_list( string $role, int $id ): array {
+		$athletes_data = [];
+		$athletes      = get_posts( [
+			'numberposts' => - 1,
+			'post_type'   => 'athlete',
+			'meta_query'  => [
+				[
+					'key'   => $role,
+					'value' => $id,
+				]
+			],
+		] );
+
+		if ( ! empty( $athletes ) ) {
+
+			foreach ( $athletes as $athlete ) {
+				$parent          = get_field( 'parent', $athlete->ID );
+				$athletes_data[] = [
+					'ID'              => $athlete->ID,
+					'first_name'      => get_field( 'first_name', $athlete->ID ),
+					'last_name'       => get_field( 'last_name', $athlete->ID ),
+					'birthday'        => get_field( 'birthday', $athlete->ID ),
+					'team'            => get_field( 'team', $athlete->ID ),
+					'parent'          => $parent,
+					'parent_name'     => $parent->first_name . ' ' . $parent->last_name,
+					'parent_id'       => $parent->ID,
+					'verified_status' => '',
+				];
+			}
+		}
+
+		return $athletes_data;
 	}
 
 	function get_athletes( WP_REST_Request $request ) {
@@ -592,48 +697,9 @@ class PKI_REST_Entities_Controller extends WP_REST_Controller {
 
 			if ( ! is_wp_error( $user ) && $user !== false && ( user_can( $user_id, 'create_team' ) || user_can( $user_id, 'create_athlete' ) ) ) {
 
-				$args = [
-					'numberposts' => - 1,
-					'post_type'   => 'athlete',
-				];
-
-				if ( $data['currentRole'] === 'coach' ) {
-					$args['meta_query'] = [
-						[
-							'key'   => 'coach',
-							'value' => $user_id,
-						]
-					];
-				} else if ( $data['currentRole'] === 'parent' ) {
-					$args['meta_query'] = [
-						[
-							'key'   => 'parent',
-							'value' => $user_id,
-						]
-					];
-				}
-
-				$athletes = get_posts( $args );
+				$athletes = $this->get_athletes_list( $user->roles[0], $user_id );
 				if ( ! empty( $athletes ) ) {
-					$athletes_data = [];
-
-					foreach ( $athletes as $athlete ) {
-						$parent          = get_field( 'parent', $athlete->ID );
-						$athletes_data[] = [
-							'ID'              => $athlete->ID,
-							'first_name'      => get_field( 'first_name', $athlete->ID ),
-							'last_name'       => get_field( 'last_name', $athlete->ID ),
-							'birthday'        => get_field( 'birthday', $athlete->ID ),
-							'team'            => get_field( 'team', $athlete->ID ),
-							'parent'          => $parent,
-							'parent_name'     => $parent->first_name . ' ' . $parent->last_name,
-							'parent_id'       => $parent->ID,
-							'verified_status' => '',
-						];
-					}
-
-
-					wp_send_json_success( $athletes_data );
+					wp_send_json_success( $athletes );
 				}
 
 				wp_send_json_error( 'Athletes not found' );
